fixed Oredict
